Move code to the hnsw index base class as we no longer depend on the float type.
